% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_df.R
\name{document_df}
\alias{document_df}
\title{Prints framework for documenting a data frame}
\usage{
document_df(df, num_spaces = 4)
}
\arguments{
\item{df}{data frame object; can be a \code{data.frame}, \code{data.table},
or \code{tibble}}

\item{num_spaces}{number of a spaces per tab; default is 4}
}
\value{
nothing
}
\description{
Prints the standard fframework for documenting a data frame in
    an R package.
}
\examples{
set.seed(0)
dat <- data.frame(x = c(LETTERS[1:5]),
                  y = c(sample(1:100, 5)))
document_df(dat)

}
