% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_component_size.R
\name{filter_component_size}
\alias{filter_component_size}
\title{Filter a tidygraph by component size}
\usage{
filter_component_size(.data, min_size = 0, max_size = Inf)
}
\arguments{
\item{.data}{a tidygraph object (for use in a pipeline)}

\item{min_size, max_size}{minumum and maximum sizes of components to keep
(inclussive bounds)}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Filter a tidygraph object by component size.
}
\examples{
gr <- tidygraph::bind_graphs(tidygraph::play_barabasi_albert(3, 2),
                             tidygraph::play_barabasi_albert(6, 2))
igraph::count_components(gr)
new_gr <- filter_component_size(gr, min_size = 4)
igraph::count_components(new_gr)

}
