% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_index.R
\name{get_node_index}
\alias{get_node_index}
\title{Get the node indices from a tidygraph}
\usage{
get_node_index(gr, expr)
}
\arguments{
\item{gr}{a tidygraph}

\item{expr}{an expression to be quoted and evaluated in the context of the
node tibble of the tidygraph object}
}
\description{
Some functions in the 'tidygraph' package require the index of a
  node. This function evalutes an input expression in the
  \code{dplyr::filter()} function in the node tibble of a tidygraph object
  and returns the indices of the nodes that pass.
}
\examples{
get_node_index(quick_barabasi(10), name == "B")
get_node_index(quick_barabasi(10), name \%in\% c("B", "C", "D"))
get_node_index(quick_barabasi(10), stringr::str_detect(name, "A|B|C"))

}
