% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Bound values to a lower and upper limit}
\usage{
minmax(x, lower, upper)
}
\arguments{
\item{x}{vector of values to limit}

\item{lower, upper}{lower and upper limits}
}
\value{
vector of the same length as \code{x}
}
\description{
Bound the values of \code{x} to lie within the upper and lower
  bounds.
}
\examples{
set.seed(0)
c <- sample(-100:100, 20)
c
minmax(c, -10, 10)

}
