% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graphs.R
\name{n_perm}
\alias{n_perm}
\title{Calculate the number of permutations}
\usage{
n_perm(n, k)
}
\arguments{
\item{n}{length of vector}

\item{k}{number of values to take at a time}
}
\value{
a numeric value or vector if input values are a vector of vectors
}
\description{
Calcualte the number of possible permutations for a vector of
  length \code{n} taking \code{k} at a time
}
\examples{
n_perm(10, 2)
n_perm(3, 4)

}
