% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique.R
\name{n_unique}
\alias{n_unique}
\title{Number of unique values}
\usage{
n_unique(x, to_unlist = FALSE)
}
\arguments{
\item{x}{input vector}

\item{to_unlist}{boolean of whether to flatten the input with
\code{unlist()}; default is \code{FALSE}}
}
\description{
Returns the number of unique values in the input vector
}
\examples{
n_unique(c(1, 2, 3, 3, 4))
n_unique(c("A", "B", "C", "D", "A"))
n_unique(list(letters[1:5], letters[1:4]))
n_unique(list(letters[1:5], letters[1:4]), to_unlist = TRUE)

}
