% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graphs.R
\name{prep_node_names}
\alias{prep_node_names}
\title{Prepare the needed number of names}
\usage{
prep_node_names(n, name = c())
}
\arguments{
\item{n}{number of names desired}

\item{name}{a vector of names; default is \code{LETTERS}}
}
\value{
a vector of \code{n} unique names
}
\description{
Ensures there are \code{n} number of names in \code{name}. If
  not, permutations of the values in \code{name} are catentated together to
  reeach the requestion amount. This function can also be used to make
  \code{n} number of unique names by passing \code{NA} to \code{name}, in
  which case the default is to use \code{LETTERS}.
}
\examples{
prep_node_names(5)
head(prep_node_names(50))

}
