% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graphs.R
\name{random_graph}
\alias{random_graph}
\alias{quick_forestfire}
\alias{quick_barabasi}
\alias{quick_graph}
\title{Create random graphs using my preferred models}
\usage{
quick_forestfire(n, name = c(), directed = FALSE)

quick_barabasi(n, name = c(), directed = FALSE)

quick_graph(n, name = c(), directed = FALSE)
}
\arguments{
\item{n}{number of nodes}

\item{name}{node names; default is \code{LETTERS}}

\item{directed}{boolean value for directed or undirected graph}
}
\value{
a tridygraph object with named nodes
}
\description{
These are light-weight wrappers around 'tidygraph' functions to
create random graphs. They return graphs with named nodes.
}
\examples{
set.seed(0)

forest_gr <- quick_forestfire(20)
forest_gr
plot(forest_gr)

barabasi_gr <- quick_barabasi(20)
barabasi_gr
plot(barabasi_gr)

}
