% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_graph_join.R
\name{recursive_graph_join}
\alias{recursive_graph_join}
\title{Recursively join a list of tidygraphs}
\usage{
recursive_graph_join(grs)
}
\arguments{
\item{grs}{a list of tidygraph objects}
}
\value{
a single tidygraph object
}
\description{
Joins the graphs into a single graph. This recursively implementes
\code{tidygraph::graph_join()}
}
\examples{
library(tidygraph)
# TODO

}
