% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette.R
\name{show_palette}
\alias{show_palette}
\title{Show a color palette in a matrix}
\usage{
show_palette(cols, num_rows = "square", label_size = 5,
  font_family = "Times", base_size = label_size, ...)
}
\arguments{
\item{cols}{The colors to present in a matrix. If the values are named, these
will be used for the cell label. If not, then the color value itself is used
for the label.}

\item{num_rows}{The number of rows to use or if the matrix should be as
square as possible.}

\item{label_size}{The size passed to \code{ggplot2::geom_text()} for the size
of the label. Default is 5.}

\item{font_family}{The font to use for thee plot. Default is "Times".}

\item{base_size}{The base font size passed to \code{theme_bw()}. The default
is to use \code{label_size}.}

\item{...}{Any other information to pass to \code{ggplot2::theme()}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function is very simillar to the \code{show_cols()} in 'scales'. The
main difference is that this function returns a true \code{ggplot2} object.
This allows for further customization.
}
\examples{
pal <- c("#3fc5f0", "#42dee1", "#6decb9", "#eef5b2")
show_palette(pal)

show_palette(pal, num_rows = 1)

names(pal) <- paste("color", 1:4)
show_palette(pal)

}
