% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_us.R
\name{str_replace_us}
\alias{str_replace_us}
\alias{str_replace_sp}
\title{Replace underscores with spaces or \emph{vice versa}}
\usage{
str_replace_us(string)

str_replace_sp(string)
}
\arguments{
\item{string}{"Input vector. Either a character vector, or something coercible to one." (from \code{?stringr::str_replace_all})}
}
\value{
"A character vector." (from \code{?stringr::str_replace_all})
}
\description{
Vectorised over string, to replace underscores with spaces or \emph{vice versa}.
}
\details{
These two functions are thin wrappers around \code{str_replace_all()} from the 'stringr' package.
They just pass a space (\code{" "}) or an underscore (\code{"_"}) to the parameters  \code{pattern} and \code{replacement}, automatically.
(Yes, I do this often enough to warrant adding two new functions to my namespace!)
}
\examples{
txt <- "This_is_a_string"
str_replace_us(txt)

txt <- "This is a string"
str_replace_sp(txt)

}
