% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_colnames.R
\name{tidy_colnames}
\alias{tidy_colnames}
\title{Easily tidy the column names of a tibble}
\usage{
tidy_colnames(data)
}
\arguments{
\item{data}{data.frame or tibble}
}
\description{
Cleans up bad data table names by removing dots ".", spaces, and
  hyphens, replacing all with an underscore "_"
}
\examples{
tib <- tibble::tibble("colname.1" = c(1:5),
                      "colname 2" = LETTERS[1:5],
                      "colname-3" = letters[26:22])
tidy_colnames(tib)


}
