% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_pull.R
\name{u_pull}
\alias{u_pull}
\title{Pull the unique values from a column of a tibble}
\usage{
u_pull(data, var = -1, sorted = FALSE, na.rm = FALSE)
}
\arguments{
\item{data}{a table of data}

\item{var}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right.
}

The default returns the last column (on the assumption that's the
column you've created most recently).

This argument is taken by expression and supports
\link[rlang:quasiquotation]{quasiquotation} (you can unquote column
names and column positions).}

\item{sorted}{(boolean) sort the output; default \code{FALSE}}

\item{na.rm}{(boolean) remove \code{NA}; default \code{FALSE}}
}
\description{
Acts just like 'pull()' from 'dplyr' except that it only
    returns the unique values.
}
\examples{
suppressPackageStartupMessages(library(dplyr))
tibble::tibble(a = c(1:3, 1:3),
               b = LETTERS[1:6]) \%>\%
    u_pull(a)

}
