% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_na.R
\name{unique_na}
\alias{unique_na}
\title{Unique and omit \code{NA}}
\usage{
unique_na(x, to_unlist = FALSE)
}
\arguments{
\item{x}{input vector}

\item{to_unlist}{boolean of whether to flatten the input with
\code{unlist()}; default is \code{FALSE}}
}
\description{
Returns the unique list of input, removing \code{NA} values, first
}
\examples{
unique_na(c(1, 2, 3, 3, 4, NA))

}
