% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsample.R
\name{vsample}
\alias{vsample}
\title{Wrapper for \code{sample()} that always assumes input is a vector of options}
\usage{
vsample(x, size = length(x), replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose, or a
positive integer. See ‘Details.’}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the
vector being sampled.}
}
\value{
For sample a vector of length size with elements drawn from either x
}
\description{
The function \code{base::sample()} accepts either a vector of valuesto sample
from, or a single numeric value that is turned into a vector from 1 to the
input value and this range is sampled from. This has resulted in a rather
frustrating (and perplexing) error for me in the past. This function,
\code{vsample()} always assumes the input is a vector. Therefore, if the
input vector is of length 1, that value is just returned (i.e. sampling 1
value from a vector of length 1 will always select the single element).
}
\examples{
sample(10)
vsample(10)

vsample(10, size = 3, replace = TRUE)

sample(c(1, 2, 3))
vsample(c(1, 2, 3))

}
