% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_graph_join.R
\name{recursive_graph_join}
\alias{recursive_graph_join}
\title{Recursively join a list of tidygraphs}
\usage{
recursive_graph_join(grs, by = "name")
}
\arguments{
\item{grs}{a list of tidygraph objects}

\item{by}{what to group by (passed to \code{tidygraph::graph_join()})}
}
\value{
a single tidygraph object
}
\description{
Joins the graphs into a single graph. This recursively implementes
\code{tidygraph::graph_join()}
}
\examples{
set.seed(0)

gr_list <- purrr::map(c(10, 15, 20), quick_forestfire)
gr <- recursive_graph_join(gr_list)
gr

plot(gr)

}
